refactor(ui): demote misclassified components
