+++ src/app/(admin)/admin/products/_components/ProductEditDialog.tsx
@@ -4,10 +4,8 @@ import type React from "react";
 import { useActionState, useEffect, useState } from "react";
 import { updateProduct } from "@/actions";
 import type { Product } from "@/services";
-import {
-  Alert,
-  Spinner,
-} from "@/ui/components/molecules";
+import { Spinner } from "@/ui/components/atoms";
+import { Alert } from "@/ui/components/molecules";
 import { ImageField, SelectField } from "@/ui/components/organisms";
 import { NumberField } from "./NumberField";
 import {
