+++ src/app/(main)/privacy/page.tsx
@@ -1,4 +1,4 @@
-import { LegalDocumentTemplate } from "@/ui/components/templates";
+import { LegalDocumentTemplate } from "@/ui/components/molecules";
 import { PRIVACY_SECTIONS, PRIVACY_EFFECTIVE_DATE } from "./_constants";
 
 // 법무 미검수 초안 — 서비스 오픈 전 검토 필요. 보호책임자 연락처는 TODO(legal).
