+++ src/ui/components/atoms/date-display.tsx
@@ -9,7 +9,7 @@ interface DateDisplayProps {
 }
 
 /**
- * 날짜를 다양한 형식으로 렌더링하는 순수 UI 컴포넌트 (Molecule)
+ * 날짜를 다양한 형식으로 렌더링하는 순수 UI 컴포넌트 (Atom)
  * 복잡한 포맷팅 로직은 utils/date.ts 로 분리되었습니다.
  */
 const DateDisplay = ({ date, type, className }: DateDisplayProps) => {
