+++ src/ui/components/atoms/index.ts
@@ -8,6 +8,7 @@ export * from "./checkbox";
 export * from "./collapsible";
 export * from "./command";
 export * from "./dialog";
+export * from "./date-display";
 export * from "./dropdown-menu";
 export * from "./input";
 export * from "./label";
@@ -20,6 +21,7 @@ export * from "./separator";
 export * from "./sheet";
 export * from "./sidebar";
 export * from "./skeleton";
+export * from "./spinner";
 export * from "./switch";
 export * from "./table";
 export * from "./tabs";
