+++ src/ui/components/molecules/ProductCard.tsx
@@ -1,13 +1,10 @@
 import { Sparkles } from "lucide-react";
 import Link from "next/link";
-import type { Product } from "@/core/domain";
+import type { Product, SubCategory } from "@/core/domain";
+import { routes, subCategoryLabels } from "@/core/domain";
 import { calculatePrice } from "@/core/utils";
-import type { SubCategory} from "@/core/domain";
-import { subCategoryLabels } from "@/core/domain";
-import { Badge } from "../atoms/badge";
-import { TypographyMuted } from "../atoms/typography";
-import { CloudImage } from "../molecules/CloudImage";
-import { routes } from "@/core/domain";
+import { Badge, TypographyMuted } from "@/ui/components/atoms";
+import { CloudImage } from "./CloudImage";
 
 export function ProductCard({ product, rank }: { product: Product; rank?: number }) {
   const finalPrice =
