+++ src/ui/components/molecules/index.ts
@@ -5,9 +5,10 @@ export * from "./BaseSelect";
 export * from "./CloudImage";
 export * from "./ComboboxField";
 export * from "./CursorPagination";
-export * from "./DateDisplay";
 export * from "./DateField";
+export * from "./LegalDocumentTemplate";
 export * from "./LoginEntryButton";
 export * from "./PageTitle";
+export * from "./Pricing";
+export * from "./ProductCard";
 export * from "./QuantityStepper";
-export * from "./Spinner";
