+++ src/ui/components/organisms/index.ts
@@ -5,6 +5,7 @@ export * from "./BottomActionBar";
 export * from "./ChangePasswordForm";
 export * from "./CheckoutForm";
 export * from "./ClipboardButton";
+export * from "./ComingSoonTemplate";
 export * from "./InvitationFormView";
 export * from "./CoupleInfoSection";
 export * from "./CreateGuestbookForm";
@@ -25,8 +26,6 @@ export * from "./PersonValueCard";
 export * from "./PaymentMethodSelector";
 export * from "./PremiumFeatureDialog";
 export * from "./PremiumFeatureRegistrationForm";
-export * from "./Pricing";
-export * from "./ProductCard";
 export * from "./ProductCatalog";
 export * from "./ProductFeatures";
 export * from "./ProductFilters";
