--- src/ui/components/templates/index.ts
@@ -1,2 +1,0 @@
-export * from "./ComingSoonTemplate";
-export * from "./LegalDocumentTemplate";
